split logic package
