+++ src/main/java/netcracker/practice/crossgen/logic/angle/DiagonalAngle.java
@@ -1,4 +1,8 @@
-package netcracker.practice.crossgen.logic.grid;
+package netcracker.practice.crossgen.logic.angle;
+
+import netcracker.practice.crossgen.logic.grid.Grid;
+import netcracker.practice.crossgen.logic.grid.MutableGrid;
+import netcracker.practice.crossgen.logic.word.Word;
 
 import java.util.List;
 import java.util.ArrayList;
