+++ src/main/java/netcracker/practice/crossgen/logic/grid/MutableGrid.java
@@ -1,5 +1,7 @@
 package netcracker.practice.crossgen.logic.grid;
 
+import netcracker.practice.crossgen.logic.word.Word;
+
 public interface MutableGrid extends Grid {
 
     void setSymbol(int row, int col, char symbol);
