+++ src/main/java/netcracker/practice/crossgen/logic/word/Word.java
@@ -1,4 +1,6 @@
-package netcracker.practice.crossgen.logic.grid;
+package netcracker.practice.crossgen.logic.word;
+
+import netcracker.practice.crossgen.logic.angle.Angle;
 
 public interface Word {
